Moved dining related templatetags to Dining
