+++ assets/templates/dining_lists/dining_slot_info.html
@@ -1,6 +1,6 @@
 {% extends 'dining_lists/dining_slot.html' %}
 {% load staticfiles %}
-{% load form_extras %}
+{% load dining_tags %}
 
 {% block details %}
     <div class="row">
